Add missing labals
